Continue dbclass.py
Finished the ThornyUser builder, all that is left is to somehow pass on the Thorny Bot OR alternatively, a user (User, Member or ctx.author) variable instead of the base `user_id` and `guild_id`. THis is so I can do operations such as fetching and updating the username.
Although, I think I could probably turn this into a method of the THornyUser class rather than make it automatic, but it would be nice to have it automatic.