Began Database Connection changes
This should fix the biggest issue which is that connections keep piling up. Rather than acquiring a Connection object during `ThornyFactory.build()` and passing that to everything, I will now be passing through a Pool object, and then within the `ThornyEvent` calls and `commit()` calls, the program uses `async with pool.acquire() as conn` which automatically releases the connection afterwards.

Additionally I created a new markdown file for the changelogs as that's way easier for me to keep track of. As long as I actually write to it.

There is a ton of work for me to continue for this big change, as it does require a significant rewrite. Any changes to the ThornyUser object must happen within it, and then followed by a call to `dbcommit.commit()`, which commits the changes to the database. Makes things 100x easier.