Done some testing to conclude the following:
Running the discord bot in the `after_server_start` function using the `create_task` coroutine works, however it causes blockages in the loop. There is something that blocks the loop, and discord cannot send the heartbeat, causing it to shut down randomly.

Check discord.log for more information on the issues.