V1.10 (#104)
* fix bugs to do with connections

* Make the Connection Logs better

* Breaking Change:
Move away from the thorny.activity table and into the new thorny.playtime table, which features a playtime_id

* Breaking change:
Channels are now dynamic, using the get_channel() method.

* Make use of the new Project Application system.

TODO:
- Make Projects commit to the db
- Add more projects features

* Use the new server status indicators

TODO:
- Make Projects commit to the db
- Add more projects features

* Start building up the Projects class.

TODO:
- Make Projects commit to the db
- Add more projects features

* Refactor Projects. Create new separate project.py file, and a new Factory for projects.

TODO:
- Make commits

* Add project commits

* Add new Project Embed
Make PersistentProjectAdminButtons work with the new Project system

* Alpha version of Projects are ready to be released.

* CHange playtime.py to match latest 4.0 development

* Finish update v1.10.

This update includes an early version of the new Project Manager. Not all features are in yet, but I want the system to work for now and will add the features later.

* Finish update v1.10.

This update includes an early version of the new Project Manager. Not all features are in yet, but I want the system to work for now and will add the features later.

* Bump version number

---------

Co-authored-by: Pavel <pavel@krutov.eu>